requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
pytest
httpx
python-dotenv
